WIP Support for Anki 2.1. Basic functions and all services already work.
This is the result of conversion with '2to3-3.6' tool, manual
refactoring of PyQt4 to PyQt5, migration to new packaging system,
migration from BeautifulSoup version 3 to version 4 and so on.

The pieces of code migrated automatically with 2to3 are closest
possible equivalents of Python2 code, but should be refactored
to really get advantage of Python3 (e.g:
    for k, v in list(d.items()):
should not create a list for iteration).